msp_automation_chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
kubernetes==28.1.0
pyyaml==6.0.1
prefect==3.4.6
prefect-kubernetes==0.3.0